\name{madlib.lda}

\alias{madlib.lda}

\title{
  Latent Dirichilet Allocation
}

\description{
  The wrapper function for MADlib Latent Dirichlet Allocation. The
  computation is parallelized by MADlib if the connected database is
  Greenplum database. 
}

\usage{
madlib.lda(data, docid, words, topic_num, alpha, eta, iter_num = 20,...)
}

\arguments{

\item{data}{
    An object of \code{db.obj} class. This is the database table
    containing the documents on which the algorithm will train.
    The text of each document should be tokenized into 'words'.
  }
  
\item{docid}{
    Text name of the column containing the id of the documents.
}

\item{words}{
    Column name of the input data table containing the vector of words/tokens in the documents.
}
  
\item{topic_num}{
    Desired number of topics
}

\item{alpha}{
  Hyperparameter for the Dirichlet prior which determines the per-document topic distribution.
}

\item{eta}{
  Hyperparameter for the Dirichlet prior which determiens the per-topic word distribution.
}

\item{iter_num}{
  Number of iterations to be run.
}
}

\value{

An object of type lda.madlib which contains the following values:

\item{assignments}{
  A list of the same length as the number of documents. Element i of this list is an integer vector
  indicating the topic assignment for each word in document i.
}

\item{topics}{
  A topic_num by vocab_size matrix where each entry indicates the number of times a given word
  was assigned to a given topic.
  
}

\item{topic_sums}{
  A topic_num by 1 matrix containing the total number of words assigned to each topic.
}

\item{document_sums}{
  A topic_num by number of documents matrix which contains the total number of words in each document
  assigned to each topic.
}

\item{model_table}{
    A db.obj pointing to the model table generated by the madlib.lda_train function.
}

\item{output_table}{
    A db.obj pointing to the output table generated by the madlib.lda_train function.

}

\item{tf_table}{
    A db.obj pointing to the output table generated by the madlib.term_frequency function.

}


}



